refactor: 파일 구조 개선 - 기능과 스타일 분리
