+++ components/common/Button/style.ts
@@ -1,16 +1,4 @@
-import React from 'react';
-import clsx from 'clsx';
-
-interface ButtonProps {
-  children?: React.ReactNode;
-  onClick?: () => void;
-  disabled?: boolean;
-  variant?: keyof typeof variantClasses;
-  size?: keyof typeof sizeClasses | string;
-  className?: string;
-}
-
-const sizeClasses = {
+const sizes = {
   button:
     'w-[109px] h-[32px] text-medium12 md:w-[72px] md:h-[30px] md:text-medium14 lg:w-[84px] lg:h-[32px] rounded-sm',
 
@@ -27,36 +15,10 @@ const sizeClasses = {
     'w-full h-[52px] text-medium16 md:w-[320px] md:h-[62px] md:text-medium18 rounded-lg',
 };
 
-const variantClasses = {
+const variants = {
   solid: 'bg-violet text-white disabled:bg-gray400',
   outline: 'bg-white border border-gray300 text-gray500',
   ghost: 'bg-white border border-gray300 text-violet',
 };
 
-export default function Button({
-  children,
-  onClick,
-  disabled = false,
-  variant = 'solid',
-  size = 'button',
-  className,
-}: ButtonProps) {
-  const selectedSize = sizeClasses[size as keyof typeof sizeClasses] ?? size;
-
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={disabled}
-      className={clsx(
-        'flex cursor-pointer items-center justify-center overflow-hidden p-2 font-medium',
-        selectedSize,
-        variantClasses[variant],
-        disabled && 'cursor-not-allowed',
-        className
-      )}
-    >
-      <div className="overflow-hidden text-ellipsis whitespace-nowrap">{children}</div>
-    </button>
-  );
-}
+export { sizes, variants };
